Take dashboard data loaders out of internal
JIRA: RAIL-2956
