+++ libs/sdk-ui-ext/src/dataLoaders/DateDatasetsDataLoader.ts
@@ -1,6 +1,7 @@
 // (C) 2021 GoodData Corporation
 import { IAnalyticalBackend, ICatalogDateDataset } from "@gooddata/sdk-backend-spi";
-import { dataLoaderAbstractFactory } from "../../../../dataLoaders/DataLoaderAbstractFactory";
+
+import { dataLoaderAbstractFactory } from "./DataLoaderAbstractFactory";
 
 interface IDateDatasetsDataLoader {
     /**
