+++ libs/sdk-ui-ext/src/dataLoaders/constants.ts
@@ -9,3 +9,8 @@ export const LOADER_CACHE_SIZE = 5;
  * How many insights will be kept for each workspace.
  */
 export const INSIGHT_CACHE_SIZE = 20;
+
+/**
+ * Data for how many dashboards will be kept for each workspace.
+ */
+export const DASHBOARD_CACHE_SIZE = 20;
