+++ libs/sdk-ui-ext/src/dataLoaders/test/dataLoaders.mock.ts
@@ -1,8 +1,8 @@
 // (C) 2020-2021 GoodData Corporation
 import noop from "lodash/noop";
-import { IWorkspaceInsightsService } from "@gooddata/sdk-backend-spi";
+import { IWorkspaceInsightsService, IWorkspaceDashboardsService } from "@gooddata/sdk-backend-spi";
 
-export const noopWorkspaceDashboardsService: IWorkspaceInsightsService = {
+export const noopWorkspaceInsightsService: IWorkspaceInsightsService = {
     createInsight: noop as any,
     deleteInsight: noop as any,
     getInsight: noop as any,
@@ -13,4 +13,26 @@ export const noopWorkspaceDashboardsService: IWorkspaceInsightsService = {
     getVisualizationClass: noop as any,
     getVisualizationClasses: noop as any,
     updateInsight: noop as any,
+};
+
+export const noopWorkspaceDashboardsService: IWorkspaceDashboardsService = {
+    createDashboard: noop as any,
+    createScheduledMail: noop as any,
+    createWidgetAlert: noop as any,
+    deleteDashboard: noop as any,
+    deleteWidgetAlert: noop as any,
+    deleteWidgetAlerts: noop as any,
+    exportDashboardToPdf: noop as any,
+    getDashboard: noop as any,
+    getDashboardWidgetAlertsForCurrentUser: noop as any,
+    getDashboardWithReferences: noop as any,
+    getAllWidgetAlertsForCurrentUser: noop as any,
+    getDashboards: noop as any,
+    getResolvedFiltersForWidget: noop as any,
+    getScheduledMailsCountForDashboard: noop as any,
+    getWidgetAlertsCountForWidgets: noop as any,
+    getWidgetReferencedObjects: noop as any,
+    updateDashboard: noop as any,
+    updateWidgetAlert: noop as any,
+    workspace: "workspace",
 };
