+++ libs/sdk-ui-ext/src/index.ts
@@ -1,4 +1,4 @@
-// (C) 2007-2020 GoodData Corporation
+// (C) 2007-2021 GoodData Corporation
 
-export { clearInsightViewCaches } from "./dataLoaders";
+export { clearInsightViewCaches, clearDashboardViewCaches } from "./dataLoaders";
 export * from "./insightView";
