--- libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/dataLoaders/constants.ts
@@ -1,6 +1,0 @@
-// (C) 2021 GoodData Corporation
-
-/**
- * Data for how many dashboards will be kept for each workspace.
- */
-export const DASHBOARD_CACHE_SIZE = 20;
